Updated Ajax, find, and removeSelf
Ajax now accepts a selector and returns the containing html from that.

Find now runs FW which will convert the object to an fw object if it is
not. This is handy for running functions inside iframes or other
objects.

The remove self was added. Simple shortcut for
elem.parentNode.removeChild(elem)