Fixed Modal Problems
I changed so that the id mapped to the class on the demo area, well the
id of the container was modal, so then I had two modals in the modal
call which died. I wonder if I really want this modal, or use array to
have an array the other would only bring back one, no I don't think so.